requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
emoji==1.5.0
idna==3.2
kss==3.2.0
pydub==0.25.1
python-dotenv==0.19.0
requests==2.26.0
urllib3==1.26.7